encodePreservingSource: ByteRange capable SPiece
For certain use-cases, it is required to track translated text to
the parts in source-text which generated the translations.

Implementation of this feature in this commit takes in a string_view for
a line as an input and provides the Words for consumption ahead in
marian, along with the byte-ranges (as string_views) in the source for
each of the words coming out of sentencepiece. Absent these ranges,
normalization leads to loss of information of the source-text which
generated the translation.

This commit includes the following additions/changes to realize this
functionality.

- encodePreservingSource(string_view &line, std::vector<string_view>
  &alignments. ...) realizes feature described. The function is added to
  IVocab interface, with aborts on implementations that do not have the
  functionality. For the time being, only SentencePieceVocab has the
  function implemented properly.

- SentencePiece now through marian-dev CMake flags
  (-D_USE_INTERNAL_STRING_VIEW) switches to absl::string_view instead of
  an earlier minimal implementation. absl::string_view (not minimal) is a
  requirement for downstream applications. absl::string_view is
  typedef-ed into marian::string_view for use ahead.

- Successful build now also requires include of SentencePiece's
  SentencePieceText definitions consequently creating a protobuf
  include dependency at compile which did not exist before.